+++ src/main/java/lotto/LottoAppMain.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.woowacourse.lotto.domain.*;
-import com.woowacourse.lotto.domain.dto.LottoRankDTO;
+import com.woowacourse.lotto.dto.LottoRankDTO;
 import com.woowacourse.lotto.domain.factory.LottosFactory;
 import com.woowacourse.lotto.view.InputViewConsole;
 import com.woowacourse.lotto.view.OutputViewConsole;
